+++ core/src/io/worxbend/gitea4s/error/GiteaError.scala
@@ -11,6 +11,7 @@ object GiteaError:
   final case class NotFound(message: String, body: String) extends GiteaError
   final case class MethodNotAllowed(message: String, body: String) extends GiteaError
   final case class Conflict(message: String, body: String) extends GiteaError
+  final case class PreconditionFailed(message: String, body: String) extends GiteaError
   final case class UnprocessableEntity(message: String, body: String) extends GiteaError
   final case class Locked(message: String, body: String) extends GiteaError
   final case class RateLimited(resetAt: Option[Instant], body: String) extends GiteaError